Gets photo, exports the outline - New
This file was created today, 4/4/2017, it gets the picture and will
create the outline of it. It will resize the picture to make it square
and will export the png along with a csv of the outline at 10, 20 ,30
,40, 50 pixel square.

It needs comments to make it more readable, also it needs a better way
to get the path of the picture.